IDamage Interface and Dammage() Feat & HUD Script
Made the Interface for receiving damage of the player, the Damage function. The player's health will start at 3 (as 3 lifes or hearts), made a delegate that will listen for functions when the player receives damage.

Also made the HUD script, added the delegate of the function onDamageRecieved so that it calls the function ModifyHealth which will modify the health bar fill and gradient or that will reduce the hearts amount.