Imported more .fbx and made their prefabs
Imported to Assets\Art\FBX\Furniture:
Botiquin/FirstAid.fbx

Imported to Assets\Art\FBX\Throwable:
Radio.fbx
MeasuringTape.fbx

Imported to Assets\Art\Textures the textures of the FirstAid, the Radio and the MeasuringTape.

Made the prefab of each fbx in its respective folder in the prefab folder with its colliders.

Also disabled the FirstAid with no material and added the point light to the FirstAid with material.