Flashlight PickUp and Behaviour Scripts
Made a new PickUp script for the flashlight which inherits from the Interactable class. The difference between the Objects Pick Up is that this can't be thrown and only dissables the flahslight in the map and enables tha flashlight in the player.

The Flashlight Behaviour script for now it only turns on and off the flashlight