Hide Ability feat
Assets\Prefabs\Furniture\Telas.prefab
Selected the purple cloth of the prefab and added a child which has a camera that will be the "hidden camera". Added in the purple cloth the DiscountCloth.cs.

DiscountCloth,cs
Made an script that inherits from Interactable that, when is able to interact, lets the player "hide", which is done by disabling the MainCamera and enabling the hidden camera.

Assets\Scripts\Interaction\Interactable.cs
Deleted some logs.