DiscountCloth.cs && Choke() Refactor // RecoverBreath() Feat
Assets\Scripts\Interaction\DiscountCloth.cs
Changed the listener location so that it starts listening to the delegate when the player interacts with the specific cloth and stops listening to it when the player exist the cloth. This was done because otherwise it would call the functions of all the 34 cloths.
Also added the StartCoroutine for Choke in the Hide function and the Recover coroutine in the ExitCloth function.

Assets\Scripts\Player\PlayerManager.cs
Made public the coroutine Choke, added a filter so that it cant take breath away unless the player is hidden.
Created another coroutine RecoverBreath that does the opposite of Choke(), adding breath to the player every 2 seconds, this checks that the player is not hidden and that it does this until the breath catches up to the maxBreath.