Audio Mixer Master - Soundtrack & SFX
Made the Master Audio Mixer and the Sondtrack and SFX subgroups.

Added an Audio Source component to the throwable objects and added the the function Play for each one on its OnCollissionEnter so that it sounds on every collision after being thrown. Each object should have its own audio clip for being thrown.