Started NPC's dialogue interaction system
Made DialogueTrigger.cs script child from Interactable so that the player can start the conversation with an NPC when its near it.

When the player starts a conversation the camera and player's movement will lock making him stay until the conversation is over.

For locking the camera:
CameraBehaviour.cs
Made a function LockOnConversation that will change the value of a new float lockCamera to 0 when the player is talking, so that it doesn't move weather it;s receiving input, or 1, which won't affect the normal input.

For locking the player movement:
PlayerMovement.cs
First made it a singleton and made a function ChangeSpeed so that when the plaer is talking the speed will be 0, otherwise it will be 1.

the bool isTalking is in PlayaerManager.cs

Added the dialogue canvas to the UI and made the connections to the name and dialogue text.