Made Ghost ability to "listen" to throwable objects
Now all throwable objects have a children object that contains the CastSound script and a big colliders, which is a trigger. This colliders is the representation of the range the object can be "listened" by the ghost.

The ghost Game Object that constains the GhostAI script also has a big collider, this is the one that "listens" and reacts to the onTriggerEnter os the CastSond script.